refactor: MessageUtil 패키지 변경
