+++ src/main/java/com/codenear/butterfly/auth/application/email/CustomUserDetailsService.java
@@ -1,7 +1,7 @@
 package com.codenear.butterfly.auth.application.email;
 
-import com.codenear.butterfly.auth.exception.message.MessageUtil;
 import com.codenear.butterfly.auth.domain.dto.CustomUserDetails;
+import com.codenear.butterfly.global.util.MessageUtil;
 import com.codenear.butterfly.member.domain.Member;
 import com.codenear.butterfly.member.domain.repository.MemberRepository;
 import lombok.RequiredArgsConstructor;
