+++ src/main/java/com/codenear/butterfly/global/util/MessageUtil.java
@@ -1,0 +1,17 @@
+package com.codenear.butterfly.global.util;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.context.MessageSource;
+import org.springframework.context.i18n.LocaleContextHolder;
+import org.springframework.stereotype.Component;
+
+@Component
+@RequiredArgsConstructor
+public class MessageUtil {
+
+    private final MessageSource messageSource;
+
+    public String getMessage(String code, Object... args) {
+        return messageSource.getMessage(code, args, LocaleContextHolder.getLocale());
+    }
+}
